[INTERNAL] sap.uxap: content fading in sap_belize_plus corrected
The ObjectPageHeader and ObjectPageHeaderContent are transparent
by design in sap_belize_plus in the context of sapUshellShell.

The transparency used to be achieved with opacity: 0.6

Issue
The transparency affects the appearance of the content,
the colors fade as well.

Solution
The opacity: 0.6 is replaced by the Less function fade(@bg_color, 40%),
which converts the hex type background color to rgba and as a result the content
color is not affected.

How to test
(1) Open any ObjectPageLayout Explored sample
(2) Add 'sapUshellShell' class to the 'body'
(3) Find a image url in google and set background: url(yourimage.png)
    to the xml.view DOM element in which the ObjectPageLayout is.

Change-Id: I4e28c1a91af89ff804fd9dd1c31645072f74d7d5
BCP: 1770337576