[INTERNAL] XMLComposite: multi-aggregation forwarding
-we demonstrate here several issues:

i) when sap.m.Select Control is used in an XMLComposite we have the
Problem that the binding is no longer there on the forwarded
Items, more precisely they are still there on the items aggregation
on the Select Control (i.e. this forwarding works) but are not present
on the items aggregation of the SelectList Control. Pls note, however,
that the Select is not *moving* these items from the Select to the
SelectList - they seem to co-exist though the details are not completely
clear. See “Ex D” in running application DeepBinding.html. Also cf.
2769348 and mails.

ii) the SelectList can only be used standalone when we modify updateItems
in SelectList and remove the line “this.destroyItems();” - otherwise
only the last item in the items aggregation will ‘survive’ in the
aggregation. See “Ex C” in running application DeepBinding.html.

iii) the final problem seems to be a general issue with recursive
forwarding of List-like controls. We see here that when we try to do this,
only the first item in the list we be forwarded to the inner control.
One level of forwarding works w/o a problem but with two levels and
above the issue occurs. Related to the restriction that the hidden
Aggregation of the XMLComposite “_content" is assumed multiple=false ?
See “Ex E” in running application DeepBinding.html and "Ex B" in
running application ForwardingTest4.html.

Change-Id: I5bf2a77ca25636c19d3e3b6ceea5b0878e0d0313