[FIX] AnalyticalTable: filter menu handling for measures corrected
The AnalyticalTable code was changed ~2 years back to remove filter menu
option for measure fields with change:
I49992ac923543b6c221266007241dfa919dc4631
That change however seems to be incorrect, as filtering measures can
work and should be applied on the aggregated data.
If such filtering is not possible the backend should set the property as
not filterable (e.g. sap:filterable=false)
The filtering provided by personalisation of SmartTable anyway shows the
property in the filter tab, if it is filterable.

BCP: 1770355530
Change-Id: I39d7ace74e0406b780e61d4ab9ee021ea90a95b8