[FIX] sap.ui.commons.TextField - paste of text with line endings on IE works.
In IE (not Edge), pasting of strings that contain line endings (CR, CRLF) was not supported,
as the browser used to cut the pasted string to one containing only the text before the first line ending.

Attaching to the onpaste event to remove the line endings fixes the issue.
The issue is only fixed for IE.

Change-Id: Ia7ba2dfa3e936aadd1313727e20e03f7008ade35
BCP: 1730434255