[FIX] jQuery.sap.loadResource: Use preloaded resources by fixing URL to name guessing
When calling "loadResource" with an "url" it will try to "guess" the
resource name by checking the registered module paths in order to make
use of the preload cache.
In some cases this didn't work as the URLs might be relative and expected to
be provided in the same relative format as the module path has been
registered.
This resulted into a network request every time a resource bundle has
been initialized, even though it has already been preloaded via the
Component preload.
By converting all relative URLs into absolute URLs this check is now
more robust.

Change-Id: I4c5892ad3e2693489588a0727fe057d6e3428842
BCP: 1770349916
BCP: 0020751295 0000401957 2017