[FIX] sap.uxap.ObjectPageLayout: Control responsive behavior fix
The problem:
ObjectPageHeader was not behaving in a responsive manner when the
control was showing long title and subtitle and action buttons on
certain page width the action buttons overlapped the subtitle.

The root cause:
This was happening when a control was nested inside a control that sets
css "white-space: nowrap" (for example sap.ui.layout.Splitter)
which break the responsive behavior of the ObjectPageHeader.

The solution:
Add reset for the CSS property "white-space" of the ObjectPageLayout
to the default value "normal"

Other:
QUnit test added

CR-Id: 002075125900002786402017

Change-Id: I9bfbfe483d0272ad9d40c9ce2832f8a54af81d25
BCP: 1780382804
(cherry picked from commit df769ee1af7f87f025e89ed0393f79c195102a3f)