[FIX] AnalyticalBinding: no duplicate unit column in spread sheet export
If an AnalyticalTable displays a measure and its associated unit the
spread sheet export would show the unit column twice. The problem was,
that the unit will always be added by the logic constructing the request
and there is no check if a column in the table already contains the
unit. The coding now checks if a column with the same name is already
containted in the list of columns for the spread sheet
Snippix #72022

PS1: downport

Change-Id: I331b0a060c9cdaece34acc7ac52b67c5ac1ba49a
BCP: 1780400724
CR-Id: 002075125900002995562017
(cherry picked from commit 7a52263bbed30609182cb9ed3a47b5e2c7ab54e4)