[INTERNAL] DemokitV2: BCP component added in the about dialog message
The problem:
In the message for internal release a BCP component was missing and
users ware forced to discover the needed component using different
channels.

The solution:
BCP component CA-UI5-CTR is added to the message having in mind that it
might be used not only to report incidents related to the DemoKit itself
but also for it's content.

Change-Id: I8ac104670ce34ab68350270cd67143c652eff7b6
BCP: 1770366602