[INTERNAL] Core: move registries from Core to Component and Element
Introduces a new mixin "ManagedObjectRegistry" that can be applied to a
ManagedObject subclass that introduces an ID scope (like Elements or
Components or Templates do). The mixin implements a registry of objects,
keyed by their ID. It allows to get a map object with all currently
known instances, to retrieving a single instance by its key, to execute
some callback on each instance, to filter all instances by a given
condition (predicate function) and to get the number of instances.

The change applies the mixin to Component and Element and removes the
previous implementation from the Core. Existing Core APIs redirect to
the functions provided by the mixin.

Existing workarounds to access the old internal registries in the Core
are replaced by the new APIs.

Change-Id: Ie8fab2fdf591a0cc6a9b8146ced5f18a6c7e672b
JIRA: CPOUIFPHOENIXCORE-2725