[INTERNAL][FIX] sap.uxap.ObjectPageLayout: Corrected restore scrollPosition
Restore scroll position didn't check if the latest saved scroll position is still valid.
However, the saved position will be invalidated if a section preceeding the selected section is no longer available (because of being deleted of hidden).
Fixed by adding a check that the restored scrollPosition is within the section that should be selected.

Change-Id: I06f974829ed0e6434b38a1e0aee307f6b49d518d
BCP: 1780236313
CR-Id: 002075125900002806312017