[INTERNAL][FIX] sap.uxap.ObjectPageLayout: Ensured cleanup of hidden tabs
When in iconTabBar mode, only one section-tab is rendered at the time.

Since the objectPage inserts the selected tab to the DOM via flush (replacing the previously rendered),
it also has to ensure cleanup of the previously rendered dom-element.

Fixed by calling the designated cleanupControlWithoutRendering function of the renderManager,
that allows to cleanup (e.g. deregister listeners attached on the dom element).

Change-Id: I434078def9968af8841ff48f03e79526034e1dd5