[FIX] sap.m.MenuItem - setting key does not throw an error anymore.
There is a sync between sap.m.MenuItem properties and internally
used  menu Item implementation properties, so when property on
the sap.m.MenuItem is set, this sync makes sure this property is
propagated to the target menu item instance.

Current target menu item implementations are sap.ui.unified.MenuItem
and sap.m.MenuListItem.

However not all properties of the sap.m.MenuItem are relevant for
target menu items. For example a 'key' (that comes from sap.ui.core.Item,
which is parent of sap.m.MenuItem)  should not be propagated to the target
items, because they don't have such property.

BCP: 002075129500004223242017
Change-Id: I5daf5830d25bc1f525dad8c72e67f3d6e5eb1fa9