[INTERNAL][FIX] sap.m.Fiori20Adapter: Corrected identification of current viewId
Backround:
0) When the adapter completes the parsing of the control tree, it fires an event with parsed details of the current view,
where the current view is a direct child of the topmost navContainer.
1) While the adapter parses the control tree, it also attaches listeners for future changes in the control tree
2) The listeners then re-trigger parsing of the changed area and fire an event with the updated parsed details of the current view

Problem:
The "current view" (from point 2 above) was taken to be the view that was identified as current at the point of attaching the listeners (see 1 above)
But IF at the point of attaching the listeners there were *NO views at all* [but only an empty navCntainer]
=> then the id of the referring view is set to NULL
=> the adapter fires an event with parsed details where current viewId is set to NULL

The bottom reason for the problem was that aCurrentViewPath array was used, where a null entry was accepred as VALID (as current id value of an existing view).
Fixed by:
- replacing the aCurrentViewPath with a single entry that only holds the value of the current viewId
- corrected the check for availability of current viewId

Change-Id: I562fc9b59afd719c5ff2483a51b9a5b563684c9c
BCP: 1780352623
(cherry picked from commit 94ac0ae13765fb3caf3926e2779d96165e88d0f7)
CR-Id: 002075125900003831522017