[FIX] sap.m.Page: Wrapper tag around the toolbars are properly changed
Depends on the aria roles set to the page header, sub-header and footer the HTML tag is changed. If no landmark info is set the HTML tag is set to header or respectively footer. In all other cases the HTML tag is changed to div and aria role is set through the landmark. In this way the generated HTML is always valid and we do not end up with a case where header tag accepts role banner or any other not valid role.

Change-Id: Iebc0c8acb04f549ea7b7b7da5757dd621c439f53
BCP: 1770116282