[FEATURE] sap.f.DynamicPageTitle: New aggregation navigationActions introduced
(1) New navigationActions aggregation is added to the DynamicPageTitle.
(2) The navigationActions can be positioned in two different areas, depending on the browser size.
(3) The breakpoint is 1280px.
(4) Considering the breakpoint, the navigationActions  are placed:
	(a) either in the main right area, right next to the actions
 	(b) or in the top-right area, above the actions.
(5) Observer tracks the visible property for both the actions and navigationActions
    to check and update the ToolbarSeparator`s visibility.
(6) The ToolbarSeparator`s visibility is updated whenever the actions and the navigationActions
    aggregations change.

Change-Id: I23d4ddd8deaf6b09cd6f95788d3ff8f353dd09ab