[INTERNAL] sap.f.DynamicPageTitle - flex-basis used for content area
Problem:
If a container with an overflow mechanism, such as for example sap.m.OverflowToolbar,
is used inside the content area of sap.f.DynamicPageTitle, items that overflow
can never go back, because the flex parent won't allocate back the space to the overflow container.

Background:
The Left(heading+snapped/expanded content) and Middle(content) areas of the sap.f.DynamicPageTitle
are children of a container with "display: flex". When the title is shrunk, the parent container's
width is redistributed between the Left and Middle areas according to their priority (flex-shrink).
However, if the title is expanded afterwards, and the Middle(content) area contained an overflow
container, this overflow container will not get its initial space back as the browser believes
that is has all necessary space already allocated.
For the browser to allocate more space, so that items that moved to the overflow can come back,
it needs a hint about the container's optimal size (contrary to its current size with partially
overflowed items).
This hint is given in the form of the "flex-basis" property, its value determined based on the
minimal width of the overflow container, needed to show all of its items.

Solution:
Whenever a container with overflow mechanism is added to the Middle(content) area of the
sap.f.DynamicPageTitle control, the title attaches an event that is fired whenever the content size
of the overflow container has changed. In the event handler, the "flex-basis" property of the Middle
area is manipulated.

Restrictions:
 - Only sap.m.OverflowToolbar is recognized as an overflow-enabled container for now.
   The event, fired by sap.m.OverflowToolbar, is still experimental, hence private.
 - Only one overflow-enabled container should be used at a time inside the Middle(content) area.
   Otherwise, the "flex-basis" value that will be set may be unreliable.
   For the moment this translates to one instance of sap.m.OverflowToolbar as the only content of
   the "content" aggregation.

BCP: 1770338615

Change-Id: I5f84f749ba8fbcb8caab4674ee9cfede58ac583a