[INTERNAL]sap.uxap.ObjectPageLayout: Moved check position
Moved the check for _bPersistHeaderInTitleArea outside the _expandCollapseHeader header function.
The check is still made before the function is called, so the logic (and execution order) is unchanged.

This is a step of the planned refactoring in order to align ObjectPage code with DynamicPage code (for the integration of DynamicPageTitle.

This particlar change is needed in order to allow later the function to be called independently of the _bPersistHeaderInTitleArea check.

Change-Id: Iebf8064f0b21f5fd4c34e01efd3c0f5168b7715f