[FIX] sap.m.Button: Hover on button inside transparent toolbar is now corrected for IE11
This fixes the hovering issue in IE11 that appears to be a browser issue
and is much more visible than the issue that the rule is trying to
address. The button background inside a toolbar will not be transparent
now but is almost the same as the background color of the toolbar on Belize.
On HC* themes this rule does not matter.

The CSS rule is actually excluded for IE only. This rule aims to fix
just partially the issue. For example it does not cover the disabled
buttons that also have the same background issue.

The control CSS requires major refactoring that is not part of this
bug fix.

Change-Id: I2ca7c110ecde10839ebeec4a951af96191ff6b9a
BCP: 1780198114