[INTERNAL] sap.uxap.ObjectPageLayout - IE11 getBoundingClientRect error fixed
Background:
IE has a bug, causing calling getBoundingClientRect for elements that don't have
a parent to throw "Unspecified error."
The error has been fixed for Edge, but still exists on IE.
The workaround is to check for "parentElement" before trying to call getBoundingClientRect.
If there is no parent element, then the height should be considered 0.
https://connect.microsoft.com/IE/feedback/details/829392/calling-getboundingclientrect-on-an-html-element-that-has-not-been-added-to-the-dom-causes-unspecified-error
BCP: 1770412704

Change-Id: I16a07f2963158a97dc53cf69f3df8fc42033a818