[INTERNAL][FIX] sap.ui.documentation: apply content density to the body
Currently, the content density classes are applied to the root xml view.
However, the controls in the static area are not styled correctly.
The solution is to set the classses to the body tag.

Changes:
Setting the content density classes to the body tag onAfterRendering
Removing of the additonal styling for the internal List and Settings dialogs
as they will be styled automatically.

Change-Id: Ide5c6ef20a6a4ea6559f0f346bc291d672fbfdc0
BCP: 1770407144