[INTERNAL] DemokitV2: Using new Core.notifyContentDensityChanged method
sap.ui.Core has a new API method notifyContentDensityChanged which
should be called by Applications when they change content density
cozy|compact so controls currently rendered by the app can react to
this changes especially if they rely on size calculations based on
available space or element size. This is now used in the DemoKit to
properly adapt itself and samples to content density changes instead
of invalidating only the sample part as it was by now. This is done also
for examples loaded in iframe like the tutorials.

Fixed a regression of detecting density mode on deep link to a
iframe sample introduced with another commit.

Change-Id: I201242b8490d803cf19b7095a1e195446255bcfd