[INTERNAL] RTA does request flex settings without parameteres
Until now rta request the settings from the fl lib by passing the component name. In case the component was requested with the version (which rta cannot determine easily at that point) a new request was sent. This is prevented by requesting the settings without a parameter. Thus returning the first parametrized request promise with the correct settings sent in the session.

Change-Id: Ie38ca9bfd1c5b8cf864bb79e630dafe351aa488e