[INTERNAL] sap.ui.fl: refactor key for view caching
the key caching was done on the level of the Change persistence. This sometimes results in an additional lookup or even a request to get this data.
The refactoring pushes this down to the sap.ui.fl.Cache.
No tests were touched because the result and the interface is untouched.

Change-Id: I7d16d049d38986672704f7f09681fa5c0da53e30