[INTERNAL] sap.ui.fl: can request changes-bundle.json of an application
Changes created in the design time can be stored with their source code.
It is assumed the building stored the data in a file called
/changes/changes-bundle.json while the namespace is relative to the
manifest and is part of the preload.
In case of debugging where the preload can be skipped the file is
requested and can lead to a 404. Without debugging such a request is
not done (no preload = no bundle).
The request for changes must result in a valid return even if no back
end can be reached for the changes or its settings if bundled changes
are present.

Change-Id: I8a53497d907bafa2a4c2b477de4ea5647a3a5760