[FIX] Table: Avoid flickering of the busy indicator on initial data load
When using the AnalyticalTable inside the SmartTable, the SmartTable
handles the busy indicator on the first data request and then gives the
responsibility back to the AnalyticalTable. Because the
AnalyticalBinding can send multiple requests, the busy indicator would
be deactivated by the SmartTable after the response of the first request
is received, just to be immediately activated again by the
AnalyticalTable, when the second request is sent.

Related to SAPUI5 runtime change:
I2d7c8f9017ce123eaef156d89e5d124f0e75a3c9

Change-Id: I21be552a5ade48e97f5e1e96bd49caea1af1ad1c
CR-Id: 002075125900003211472017
BCP: 0020751295 0000428760 2017