[FIX] sap.m.Popover: Show scrollbar when overflowing and out of visible area.
There's some edge case for the popover where the scrollbar is not visible.
This happens when there's a content which overflows in width and height
inside the popover. In this case it was a long list with long texts
in some of the elements. When the Popover is with given width and it's
positioned at the rightmost of the screen, the scrollbar goes beyond the
window and is not visible/usable.
The issue is reproducible only on LTR. RTL is working fine.

BCP: 1780323187

Change-Id: I17c4fe813e74b8a21a3245460fb0a0f9e66df24b
CR-Id: 002075125900003940952017