[FEATURE] PlanningCalendar-accessibility for special dates & appointments
Till now, accessibility for PlanningCalendar, related to any connected
legend was poor - supported was Day, 1Week & 1Month views only if the
associated legend is rendered in the DOM. Also with recently introduced
PlanningCalendarLegend, the latter behavior is broken, because there
are 2 calendar legend sections now, but any accessibility link (describedBy,
labelledby) was always pointing one ("Calendar") section in the legend.
We have "Appointments" section that should be used if an appointment
is given with a type.

With this change the following is enabled:
- special dates with type are accessibility connected to the 'Calendar'
section in the legend if such
- appointments with a type are accessibility connected to the 'Appointments'
section in the legend if such
- if there is no legend item for given type, its type default text is used.
- all above is enabled for all built-in views (Hours, Days, 1Week, 1Month
 and Months)
- standalone usage of sap.ui.unified.Calendar & sap.ui.unified.calendar.Month
 are also fixed.

JIRA: BGSOFUIBALKAN-3495

Change-Id: Ifd28fe36e49c91ce7371a1ebc3a5d27958a73af9