[INTERNAL][FIX] sap.uxap.ObjectPageLayout: selectedSection cleared if no sections
Issue
when removeAllSection is called, the layout re-renders and tries to
compute the new scroll position, retrieving positionTopMobile of already removed selected section.

Background
There is already a method (code), which cares for invalid section selection,
called in _adjustSelectedSectionByUXRules.
The logic correctly detects that the section is not valid.
However, as this._oFirstVisibleSection is null too - no action is taken.
Then, _computeScrollPosition tries to find this._oSectionInfo[sId].positionTopMobile,
which throws a JS error as this._oSectionInfo[sId] is undefined.
Furthermore, the selectedSection continue pointing to the removed section
or the destroyed one (if destroySections is called).

Solution
I added an early return in the _adjustSelectedSectionByUXRules method,
checking if there are no sections and setting selectedSection to null.
Clearing the selectedSection prevents from further computation of the scroll position and the selected section.

BCP: 1770389068
CR-Id: 002075125900003282452017

Change-Id: Ibc96947c52798b7fdc1872be9ee26cbc4bfb55f8