[FEATURE] sap.m.SegmentedButton: new selectionChange event and new selectedItem association added
After the deprecation of the buttons aggregation both the select event and
selectedButton association are not so useful regarding the items aggregation
which is recommended to be used instead. That's why a new event and
association are introduced regarding the items aggregation while the
others are deprecated.

Explored app sample enhanced.

BCP: 1780176567
Fixes: https://github.com/SAP/openui5/issues/1514
JIRA: BGSOFUIBALKAN-3559

Change-Id: I6b58fa4358d5cd588a1e5c37336cde88d8546d91