[INTERNAL][FIX] sap.m.Select: min-width added
We add min-width to fix the issue with overlapping of the label
by the dropdown arrow.

The min-width is added,
when autoAdjustWidth = true, width = 'auto'
or in case of percentage width, such as '100%', "25%" and so on.

So, the min-width will not be added if fixed width is used,
such as '4px' or '12rem'.

The min-width value is increased with 5.125rem
to allow one letter, ellipsis and a dropdown icon fit in.

Note: If the percentage width is evaluated to less than 5.125rem,
the width of the Select will be still 5.125rem.

BCP: 1770365762
CR-Id: 002075125900003303452017

Change-Id: Idbe5c0b99773eaceeb6f11b2967cb6a940286d29