[INTERNAL][FIX] sap.m.Select: min-width for percentage width added
We add min-width to fix the issue with overlapping of the label
by the dropdown arrow.

The min-width used to be added, when
autoAdjustWidth = true or width = 'auto'.
The change adds one more case - percentage width,
such as '100%', "25%" and so on.

So, the min-width will not be added if fixed width is used,
such as '4px' or '12rem'.

The min-width value is increased with 0.125rem (5.125rem)
to allow one letter, ellipsis and a dropdown icon fit in.

Note: If the percentage width is evaluated to less than 5.125rem,
the width of the Select will be still 5.125rem.

Try out with snippix: #99003

Change-Id: I797fc309f6b53072e37898c93cba72a0dc031fb2
BCP: 1770365762