[INTERNAL][FIX] sap.f.DynamicPage: Added stacking context content
Issue:
Popover, opened within the DynamicPageHeader
is overlapped by elements in the content with big z-index.

Solution
Setting z-index 0 to sapFDynamicPageContent.

Explanation
This way, even if an element inside sapFDynamicPageContent has a big z-index,
it will not overlap anything from the DynamicPageHeader.
Because, the elements are in their own stacking context:
the elements in sapFDynamicPageContent (z-index=0) are fighting with
the elements in sapFDynamicPageHeader (z-index=2) and obviously the elements in sapFDynamicPageHeader have predominance.

The stacking context of the whole DynamicPage will look like:
sapFDynamicPage has z-index: 0
  > sapFDynamicPageTitle has z-index 2
  > sapFDynamicPageHeader has z-index 2
  > sapFDynamicPageContent has z-index 0

BCP: 1780164969
CR-Id: 002075125900003333162017

Change-Id: I0eecd91f84dea895fb9f55be5daee3d46e001cb9