[INTERNAL] DemoKitV2: API Reference fix header link to documentation
The problem:
Link to documentation in header is not visible for class that includes
one/

The root cause:
The list with the references are in the constructor node of the api.json
file not on the root node of the class.

The solution:
Get the references list for pre-processing from the constructor node.

Change-Id: Idf93917a3634bbcf39dcb3cd810ebd5b096c90bf
JIRA: BGSOFUIPIRIN-1454