[FIX] sap.m.GrowingEnablement: Fix updateDiff handling for empty diffs
In the JSON model it is possible, that an empty diff is returned (that
means all returned entries have the same property values as before), but
the contexts have been changed, because they are index based. So even if
an empty diff is returned, the context needs to be updated on all
entries of the aggregation.

BCP: 1780404339
BCP: 1770439815
BCP: 1770444014

Change-Id: I4c94a13e27caf6c30ecc8ebe87c4a0dfe3c815d1