[INTERNAL][FIX] f.DynamicPage: snapping adjustment skipped if control has no size
The DynamicPage can be used in NavContainer.
In the process of navigation,
the handlers _onChildControlsHeightChange and _onResize
reacts to the corresponding events.

At this time the DynamicPage has no width and height,
which leads to some wrong caluclations and as a result
the _adjustSnap method hides the DynamicPageHeader via _toggleHeaderVisibility(false).

Solution
Stop reacting to these events, if the control does not have size.

Change-Id: I8b8e97e59b20abdd6e78376089b2f9c0474880be
BCP: 1780274098