[INTERNAL] sap.ui.fl.support diagnostics tool contains a tree of changes
The diagnostics tool showed a flat list of changes for multiple applications.
Now the changes can be requested for a single application.
Afterwards these changes are ordered by control on which they are applied.
Additionally the data contains now information about the state of the change
and a hierarchy of dependent changes.

Change-Id: I29b3d6d770a06f9d67989d4cfc4112ee8b824a06