[INTERNAL] sap.ui.documentation - release notes are now part of the app
The release notes used to be integrated via an IFrame from a separate index.
The content of this index is now split into view, controller and css parts.
Additionally, the app's library (sap.ui.documentation.library) is used
instead of the demokit project's library (sap.ui.demokit.library).
One final addition is the introduction of busy indicator for the page.
Note: the controller code had to be refactored to pass eslint checks.

Change-Id: Iebc571a5de98e0d8b1829651287a5bff991307cd