[INTERNAL] Core: allow configuration of debug mode via bootstrap tag
Allows to configure the debug mode, esp. the partial debug mode, via an
attribute of the bootstrap script tag. 

For productive code this is of no relevance, but test code or sample
pages can use it for illustration purposes or to load parts of the UI5
code from alternative locations, while the majority of modules is still
loaded from standard library bundles.

Change-Id: I8f477ebfde68410867e3bbad718ec4e55493c5e6