[INTERNAL][FIX] sap.f.DynamicPage: Ensured fitContent check is made
The "sapFDynamicPageContentFitContainer" class must be set on the DynamicPage whenever one of the following is the case:
- no scrollable content
- fitContent enabled

However, when the first condition (no scrollble content) is the case and the page *is rerendered*, the "sapFDynamicPageContentFitContainer" class was not set,
because the check for necessity of the class was made only if content size changed.

Fixed by ensuring that the check [for the need of "sapFDynamicPageContentFitContainer" class] is always made on after rendering.

Change-Id: Ic0877f10a58e09566ee7ec445e9f23e531c9f1ce
BCP: 1770175158