[FIX] sap.ui.layout.Grid: Filter for DOM visibility added
Right now, when the ListsItemBase generates accessibility
text for a given control, it only checks if that control is
visible by using the getVisible() method. Unfortunatelly,
there are some controls using CSS to hide themselves.

sap.ui.layout.Grid is one of those. It's items are hidden
with CSS classes, rather than using the visible property.
This way getVisible() returns true even for the hidden
elements, which makes them readable for screen readers.

Due to this, a new filter has been added in the
getAccessibilityInfo method of the Grid, which simply checks
if the content is also visible in the DOM.

Change-Id: I8142b3fa5994d7d6a11bc22c5c6f283701b2cb78
BCP: 1780166841