[INTERNAL] DemokitV2: Instrumenting api-index json files
- Follow up from Gerrit change: 2772035

- api-index.json is loaded only instead of all api.json file and
navigation tree in API reference section is created from it eliminating
a lot of network requests and on route match having the info from the
index the correct api.json file for the requested class is loaded.

- api-index-deprecated.json and api-index-experimental.json are
respectively loaded when deprecated or experimental sections are loaded
and collecting their data at runtime is removed.

- libsData model is removed and alternative entity data is now
instrumented only in API reference. Obsolete code relaying on
all api.json data to be loaded at the same time is removed.

Change-Id: I9f23cbed633c52760e1378dfc3ecd29ef809277d
JIRA: BGSOFUIPIRIN-2063
JIRA: BGSOFUIPIRIN-1945