[INTERNAL][FIX] sap.uxap.ObjectPageLayout: Added safe-check for headerContent
The headerContent is internally created onBeforeRendering (and its creation depends on the specified headetTitle type).

However, if the code tries to access headerContent before the point of its creation => an error is thrown (see example situation in the supplied qunit).

Fixed by adding the missing save-check for existence of headerContent.

Change-Id: Ib122be68f477b72c31cddbddcde2f98a1b1d5403