[FIX] ManagedObject: Remove null context propagation
In case the same context was returned again by a ContextBinding, the
change handler in ManagedObject did first set and propagate a null
context, to make sure all dependent bindings will be refreshed properly.
This change introduces a force update flag on the Context, which can be
set by the ContextBinding to achieve a refresh of dependent bindings.
ManagedObject does no longer set the context to null, but instead checks
whether the context is flagged as force update and sets/propagates it,
even if it is the same context instance as before.

PS3: adapt sap.ui.model.odata.v4.ODataContextBinding#execute
PS4: JIRA added

Change-Id: I2273c03745078a660611ede8bd890b428bf3c6cd
JIRA: CPOUI5UISERVICESV3-906
BCP: n/a
CR-Id: 002075125900004069762017