[FEATURE] sap.m.Table: popinLayout property added
This property provides an enumerated pop-in row layout options. Using
this property, the Table can be configured to render the pop-in rows
with a grid layout. This property allows the pop-in rows to occupy less
space and as a result, more content will be displayed on the screen.

Resolves:
JIRA: CPOUIFTEAMB-841

Change-Id: Ie1f4e84a6ea600c751c0dce972ce18a7b0daba9b