[INTERNAL] sap.uxap.ObjectPageLayout - object page resize is propagated to the dynamic header
When dynamic page is resized, it notifies the dynamic page title.
The same should be done in object page when it's using the dynamic page header.
In addition, some testsuite samples were corrected so that the resize handler
can work correctly with them (ResizeHandler.register cannot be run before the
core is ready, so attachInit was added).

Change-Id: I48437dfeddf347a773316714b4a4984fc002ad5a