[INTERNAL] sap.f.DynamicPageTitle - content flex-basis survives re-rendering
If an overflow toolbar is used in the content aggregation, it reports its
optimal width to the dynamic page title and it sets it as flex-basis.
However, if the title is rerendered, this flex-basis is lost as it is
an inline style.
This change makes the flex-basis value part of the dynamic page title
state. It is maintained in a private variable and passed to the renderer.

Change-Id: I8531a0f5ae2427ea18ca97dbc7bc43e6aeeebeb5