[INTERNAL][FIX] sap.ui.fl.support diagnostics tool react on debugging
The sap.ui.fl diagnostics tool did not react correctly in case the debugging was not active and data was requested.
This fix adds a check to ensure the fl library is debugged.
If this is not the case data was not collected and no data is shown in the view.
To hint the user that he has to enable the debug mode a info text is shown how to enable this view.

Change-Id: I642ef51ee62791774d0a865f6c5f65dc0dab9e60
BCP: 1770504904
CR-Id: 002075125900003814992017