[INTERNAL] sap.f.DynamicPage: actions area size increased
The actios area width is changed from 17% to 20%
for phone and tablet sizes.

At first,
the visual design suggested one action (button) to remain always visible (out of the overflow).
It turns out that there are floorplans, in which a SegmentedButton with two buttons
is oftenly used. In this case the SegmentedButton would be cut.

We (VD + GD + Development) decided to increase the width of the area.

Change-Id: I032b5946f890d253e1493ec43d128035c61fcbd0