[INTERNAL] sap.f.DynamicPageTitle - CSS context added for the heading area
The DynamicPageTitle heading aggregation is expected to aggregate sap.m.Title.
However, margin-top should be applied to the Title to fulfiil the Fiori2.0 VD.

Until now, the re-styling (mentioned above) was part of the DynamicPageTitle.less.
Now we are moving to the more general solution by providing a context class and
delegate the required re-styling to the control in use.

Note: only top-level controls from type sap.m.Title,
used in the DynamicPageTitle heading aggr will be re-styled.

Change-Id: Idbfefe75e25063d3435b6b4d4d8b83cdac5bb371