[INTERNAL] sap.f.DynamicPageTitle - expandedHeading and snappedHeading aggregations added
The two new aggregations work similarly to "expandedContent" and "snappedContent"
(they switch when the header snaps/expands). The old "heading" aggregation is
not deprecated, but rather works as a replacement for both.
To achieve the new functionality the dynamic page/object page were slightly
reworked to call a single title function that takes care of aggregations internally.
Additionally, the testsuite samples were adapted.

Change-Id: I1a50b07f9d6e436877ad9f39e79a25175f2f61fe