[INTERNAL] DemokitV2: Borrowed methods fix
The problem:
If we have a symbol in the inheritance chain that has no documentation
the script that discovers the chain goes in to infinity loop.

The solution:
If a symbol name is not found in the available symbol list during
inheritance chain discovery we break the whole operation and display the
result up to the symbol not found.

CR-Id: 002075125900003787552017

Change-Id: I96b3f738c97b078afe633e1dbb9f1a72f6de87b8
BCP: 1770492427
(cherry picked from commit 68ddb1b484338c07920901ef6b84284ca6045a85)