[INTERNAL] sap.m.RatingIndicator: css fixes
BCP: n/a

The change is made in order to update control up to the visual design. 
This is needed because of the new icon style is added and its usage in some of the states and themes. 
I made the following changes:
Change the icon of DisplayOnly mode
Change the margin in list context
Fix icons and icons colors in HCB and HCW

Change-Id: Ie268dbf6cdec4c575a0b9b343daf8a36caf11e2c
CR-Id: 002075125900004107862017
(cherry picked from commit cd73b93d573ded9d5405d1507b27c8ba47909124)