[INTERNAL] sap.ui.support: Added rule for the size limit of sap.m.Select items model
By default, the sap.ui.model.Model has a default limit of 100 for the size of the bound items.

Added a rule that warns about the limit existence in one of the cases (sap.m.Select)
that generated the most support tickets from applications on the effect of the limit.

Change-Id: I4d97908610d41b6945193fd9fb0ff64698a038c0