[INTERNAL] sap.uxap.ObjectPagelayout: container size based styling enabled
Updates the media style class of the ObjectPageLayout, based on its own width,
not on the entire screen size (which media query does).
This is necessary, because the control will be embedded in other controls (sap.f.FlexibleColumnLayout for example),
and will not use the full screen width, but despite that the paddings need to be appropriate.

Example:
Consider the ObjectPageLayout is rendered in phone size ocntainer on desktop.
In this case, the phone specific styling should be applied, although the control is used on desktop.

Note:
The method is called,
when the new headers (ObjectPageDynamicPageHeaderTitle/Content controls) are used.

Change-Id: I7f1652c39ef771bc35df979a09f0c8fa2d070b57