[INTERNAL] sap.ui.support: Added rule for sap.m.Image density awareness
The sap.m.Image has density awareness enabled by default, which causes extra network requests to get the density perfect version of the image.
This hit performance if the corresponding density versions of the image do not exist on the backend server.
Added the rule to warn about the impact of the default enabling of the 'densityAware' property.

Change-Id: Id8747905cf740a14f1c7905a03cf9f507330b6ee