[INTERNAL][FIX] sap.uxap.ObjectPageLayout: Refactored scrollToSection when 0 duration
Problem:
When the page is scrolled with 0 delay, the function that toggles the anchorBar position (removes/adds the anchorBar from/to the scrollable area ) is called *after* the scroll completed.
=> Therefore the anchorBar was removed from the scrollable area *after* the page arrived at correct scroll position
=> extra pixels were scrolled (with y === anchorBar height).

Fixed by ensuring the scroll position for a section includes the anchorBar height.

Note: Qunits for the same problem were already added in the original change I561fcfd178e84af6ff38715316f4850495b7688f

Change-Id: Ie50258fe627c4dae217ba7cfdb4c2446060c72c9